requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-ckeditor==6.0.0
django-extensions==3.1.0
django-haystack==3.0
django-js-asset==1.2.2
django-simpleui==2021.1.1
django-werkzeug-debugger-runserver==0.3.1
djangorestframework==3.12.2
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.1
mysqlclient==2.0.2
oauthlib==3.1.0
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
Werkzeug==1.0.1
Whoosh==2.7.4
zope.interface==5.2.0
